Added parseNaturalLanguageDuration to core/apps
