+++ packages/core/src/v3/apps/index.ts
@@ -6,3 +6,4 @@ export * from "./http.js";
 export * from "./provider.js";
 export * from "./isExecaChildProcess.js";
 export * from "./friendlyId.js";
+export * from "./duration.js";
